More than twice as fast. Unit uses mostly ints instead of doubles now.

+++ src/ActionTypeData.h
@@ -20,7 +20,7 @@ struct ActionTypeData
     int                         buildTime       = 0;
     int                         numProduced     = 1;
     int                         startingEnergy  = 0;
-    double						maxEnergy       = 0;
+    int						    maxEnergy       = 0;
     bool                        isUnit          = false;
     bool                        isUpgrade       = false;
     bool                        isAbility       = false;
